+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 # Changelog
 
-TODO: You can follow this format for your changelog: <https://keepachangelog.com/en/1.1.0/>
+## Version 1
+### added:
+- Sticky Plaster
+- Can O' Beans
+- Chrisps x 2 flavours
+- Climber's Chalk
+- Extra Extreme Energy Drink
+- Gob Stopper
+- Icecream
+- Melon